+++ src/app/services/filters.service.ts
@@ -12,10 +12,16 @@ import 'leaflet';
     providedIn: 'root',
 })
 export class FiltersService {
+    private resultsPanelState = new BehaviorSubject(false);
+    public resultsPanelOpen = this.resultsPanelState.asObservable();
     constructor() {}
 
     private filteredSites = new BehaviorSubject<any>([]);
     selectedSites = this.filteredSites.asObservable();
+
+    changeResultsPanelState(state: boolean) {
+        this.resultsPanelState.next(state);
+    }
 
     updateSites(sites: any) {
         this.filteredSites.next(sites);